own component for the layout for the navigation
