+++ docs-next/src/components/Navigation/index.ts
@@ -1,1 +1,2 @@
 export * from './Navigation';
+export * from './NavigationLayout';
